+++ src/Components/SimpleCatalogViewer/SimpleCatalogViewer.tsx
@@ -1,14 +1,23 @@
 import React, { useEffect, useState } from "react";
 import { AxiosResponse } from "axios";
-import { Box, Cesium3DTileset, CesiumMap, CesiumResource, CesiumSceneMode } from "@map-colonies/react-components";
+import {
+    Cesium3DTileset,
+    CesiumColor,
+    CesiumConstantProperty,
+    CesiumGeojsonLayer,
+    CesiumMap,
+    CesiumSceneMode
+} from "@map-colonies/react-components";
 import { requestHandlerWithToken } from "../../Utils/RequestHandler";
 import { getRecordsQueryByID, parseQueryResults } from "../../Utils/CswQueryBuilder";
+import { validateIDsQuery, validatePositionQuery, validateShowExtent } from "../../Utils/ValidateQueryParams";
 import appConfig from "../../Utils/Config";
 import Terrain from "../Terrain/Terrain";
 import { useQueryParams } from "../../Hooks/useQueryParams";
+import { getFootprintsCollection } from "../../Utils";
 
 import "./SimpleCatalogViewer.css";
-import { validateIDsQuery, validatePositionQuery } from "../../Utils/ValidateQueryParams";
+
 
 const MAXIMUM_SCREEN_SPACE_ERROR = 5;
 const CULL_REQUESTS_WHILE_MOVING_MULTIPLIER = 120;
@@ -19,22 +28,21 @@ interface IClientFlyToPosition {
 
 const SimpleCatalogViewer: React.FC = (): JSX.Element => {
     const [models, setModels] = useState<Record<string, unknown>[]>([]);
-    const [isLoading, setIsLoading] = useState(true);
-    const [dialogErrors, setDialogErrors] = useState<string[]>([]);
+    const [footprints, setFootprints] = useState<Record<string, unknown>[]>([]);
+    // const [isLoading, setIsLoading] = useState(true);
     const queryParams = useQueryParams();
 
     let clientPosition: IClientFlyToPosition | undefined = undefined;
     let modelIds: string[] = [];
+    let shouldShowExtent = false;
     let idQueried: string | null = queryParams.get("model_ids");
     if (idQueried == null) {
-        console.error({ msg: `didn't provide models_ids` });
-        alert(
+        console.error(
             `Error: model_ids does not exists in the query params!\nA good example: "http://url?model_ids=ID1,ID2"`
         );
     } else {
         if (!validateIDsQuery(idQueried)) {
-            console.error({ msg: `models_ids param is not according to the specifications` });
-            alert(
+            console.error(
                 `Error: model_ids does not fit the specification!\nA good example: "http://url?model_ids=ID1,ID2"`
             );
         } else {
@@ -48,8 +56,7 @@ const SimpleCatalogViewer: React.FC = (): JSX.Element => {
     const positionQueried: string | null = queryParams.get("position");
     if (positionQueried != null) {
         if (!validatePositionQuery(positionQueried)) {
-            console.error({ msg: `position param is not according to the specifications` });
-            alert(
+            console.error(
                 `Error: position does not fit the specification!\nA good example: "http://url?position=lon,lat,height"\nP.S\nThe position param is optional`
             );
         } else {
@@ -62,63 +69,72 @@ const SimpleCatalogViewer: React.FC = (): JSX.Element => {
     }
     const userToken = queryParams.get("token");
     if (userToken === null) {
-        console.error({ msg: `didn't provide token` });
-        alert(
+        console.error(
             `Error: No token was provided. The token should be as a query param with the name "token".\nA good example: "http://url?token=TOKEN"`
         );
     }
 
+    const shouldShowExtentQueried = queryParams.get("show_extent");
+    if (shouldShowExtentQueried != null) {
+        if (!validateShowExtent(shouldShowExtentQueried)) {
+            console.error(
+                `Error: show_extent parameter does not fit the specification!\nA good example: http://url?show_extent=true"`
+            );
+        } else {
+            shouldShowExtent = JSON.parse(shouldShowExtentQueried);
+        }
+    }
+
     useEffect(() => {
-        if(modelIds.length > 2) {
-        //   setDialogErrors((currentErrors) => [...currentErrors, "Warning: You provided more than 2 models. This is not recommended"]);
-        console.warn("You provided more than 2 models. This is not recommended");
+        if (modelIds.length > 2) {
+            console.warn("You provided more than 2 models. This is not recommended");
         }
 
-        if(userToken) {
-          const cswRequestHandler = async (
-              url: string,
-              method: string,
-              params: Record<string, unknown>
-          ): Promise<AxiosResponse> =>
-              requestHandlerWithToken(appConfig.csw3dUrl, method, params, userToken);
-  
-          cswRequestHandler(appConfig.csw3dUrl, "POST", {
-              data: getRecordsQueryByID(modelIds, "http://schema.mapcolonies.com/3d")
-          })
-              .then((res) => {
-                  let modelsResponse = parseQueryResults(res.data, "mc:MC3DRecord");
-                  if (modelsResponse !== null) {
-                      setModels(modelsResponse);
-                  } else {
-                    setDialogErrors((currentErrors) => [...currentErrors, "Didn't find matched IDs!"]);
-                  }
-              })
-              .catch((e) => {
-                  console.error(e);
-                  setDialogErrors((currentErrors) => [...currentErrors, e.message]);
-              })
-              .finally(() => {
-                  setIsLoading(false);
-              });
+        if (userToken) {
+            const cswRequestHandler = async (
+                url: string,
+                method: string,
+                params: Record<string, unknown>
+            ): Promise<AxiosResponse> =>
+                requestHandlerWithToken(
+                    appConfig.csw3dUrl,
+                    method,
+                    params,
+                    userToken
+                );
+
+            cswRequestHandler(appConfig.csw3dUrl, "POST", {
+                data: getRecordsQueryByID(modelIds, "http://schema.mapcolonies.com/3d")
+            })
+                .then((res) => {
+                    let modelsResponse = parseQueryResults(res.data, "mc:MC3DRecord");
+                    if (modelsResponse !== null) {
+                        setModels(modelsResponse);
+                        setFootprints(
+                            modelsResponse.map((model) =>
+                                JSON.parse(model["mc:footprint"] as string)
+                            )
+                        );
+                    }
+                })
+                .catch((e) => {
+                    console.error(e);
+                });
+            //   .finally(() => {
+            //       setIsLoading(false);
+            //   });
         }
     }, []);
 
     return (
         <>
             <CesiumMap
-                center={clientPosition?.center ?? appConfig.mapCenter}
+                center={clientPosition?.center ?? JSON.parse(appConfig.mapCenter)}
                 zoom={clientPosition?.zoom ?? appConfig.mapZoom}
                 sceneModes={[CesiumSceneMode.SCENE3D]}
                 baseMaps={appConfig.baseMaps}
-                className={`simpleViewer ${isLoading ? "loading" : ""}`}
+                // className={`simpleViewer ${isLoading ? "loading" : ""}`}
             >
-              <Box className="errors">
-                {dialogErrors.map((error, i) => (
-                    <Box key={error} className="Dialog">
-                        {error}
-                    </Box>
-                ))}
-              </Box>
                 {models.map((model) => {
                     let links = model["mc:links"] as any;
                     if (Array.isArray(links)) {
@@ -147,6 +163,23 @@ const SimpleCatalogViewer: React.FC = (): JSX.Element => {
                         />
                     );
                 })}
+
+                {shouldShowExtent && (
+                    <CesiumGeojsonLayer
+                        clampToGround={true}
+                        data={getFootprintsCollection(footprints, true)}
+                        onLoad={(geojsonDataSource) => {
+                            geojsonDataSource.entities.values.forEach((item) => {
+                                if (item.polyline) {
+                                    const color = CesiumColor.CYAN;
+                                    (item.polyline.width as CesiumConstantProperty).setValue(5);
+                                    // @ts-ignore
+                                    item.polyline.material = color;
+                                }
+                            });
+                        }}
+                    />
+                )}
                 <Terrain />
             </CesiumMap>
         </>